Senha criptografada

+++ criandoapi/criandoapi/src/main/java/br/com/seteu/criandoapi/service/UsuarioService.java
@@ -2,17 +2,27 @@ package br.com.seteu.criandoapi.service;
 
 import br.com.seteu.criandoapi.model.Usuario;
 import br.com.seteu.criandoapi.repository.IUsuario;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
+
+import static sun.security.ssl.SSLLogger.info;
 
 @Service
 public class UsuarioService {
 
-    private IUsuario repository;
+    private final IUsuario repository;
+    private final PasswordEncoder passwordEncoder;
+    private static final Logger logger = LoggerFactory.getLogger(UsuarioService.class);
 
     public UsuarioService(IUsuario repository){
         this.repository = repository;
+        this.passwordEncoder = new BCryptPasswordEncoder();
     }
 
     public List<Usuario> ListarUsuario(){
@@ -21,18 +31,36 @@ public class UsuarioService {
     }
 
     public Usuario criarUsuario(Usuario usuario){
+        String encoder = this.passwordEncoder.encode(usuario.getSenha());
+        usuario.setSenha(encoder);
+        logger.info("senha criptografada: {}", encoder);
         Usuario usuarioNovo = repository.save(usuario);
         return usuarioNovo;
     }
 
     public Usuario editarUsuario(Usuario usuario){
+        String encoder = this.passwordEncoder.encode(usuario.getSenha());
+        usuario.setSenha(encoder);
+        logger.info("senha criptografada: {}", encoder);
         Usuario usuarioNovo = repository.save(usuario);
         return usuarioNovo;
     }
 
-    public Boolean excluirUsuario(Integer id){
-        repository.deleteById(id);
-        return true;
+    public Boolean excluirUsuario(Integer id) {
+        Optional<Usuario> usuario = repository.findById(id);
+        if (usuario.isPresent()) {
+            repository.deleteById(id);
+            logger.info("Usuário com ID {} excluído.", id);
+            return true;
+        } else {
+            logger.warn("Usuário com ID {} não encontrado.", id);
+            return false;
+        }
     }
 
+    public Boolean validarSenha(Usuario usuario) {
+        String senha = repository.getById(usuario.getId()).getSenha();
+        Boolean valid = passwordEncoder.matches(usuario.getSenha(), senha);
+        return valid;
+    }
 }
